+++ src/Krystal/Application/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added `createMapper()` shortcut methods in `AbstractModule` and `AbstractController`
  * Added `translateArray()` shortcut in `\View\ViewManager`
  * Added shared `createValidator()` shortcut for controllers
  * Dropped own autoloading support. Since now autoloading is managed via Composer